Split auth refresh telemetry helper
