+++ services/auth/session-v2.utils.ts
@@ -1,8 +1,6 @@
 import { Capacitor } from "@capacitor/core";
 import type { ApiSessionNonceResponse } from "@/generated/models/ApiSessionNonceResponse";
 import { commonApiFetch, commonApiPost } from "@/services/api/common-api";
-import { bucketMs } from "@/utils/monitoring/mobileLaunchTimingBuckets";
-import * as Sentry from "@sentry/nextjs";
 import { getWalletAddress, setAuthJwt } from "./auth.utils";
 import {
   getNativeRefreshToken,
@@ -10,6 +8,13 @@ import {
   removeNativeRefreshToken,
   setNativeRefreshToken,
 } from "./native-refresh-token-storage";
+import {
+  getSessionRefreshTelemetryTimestamp,
+  isUnauthorizedSessionRefreshError,
+  recordSessionRefreshFailureOutcome,
+  recordSessionRefreshOutcome,
+  withSessionRefreshAbortTelemetry,
+} from "./session-refresh-telemetry.utils";
 
 type AuthSessionClientType = "web" | "native" | "desktop";
 type RefreshTokenSessionClientType = Exclude<AuthSessionClientType, "web">;
@@ -56,30 +61,6 @@ type SessionRefreshInFlight = {
   readonly promise: Promise<SessionRefreshResponse | null>;
   activeConsumers: number;
 };
-type SessionRefreshTelemetryOutcome =
-  | "started"
-  | "success"
-  | "unauthorized"
-  | "aborted"
-  | "network_error"
-  | "backend_error"
-  | "cooldown_used_empty"
-  | "cooldown_used_retry"
-  | "deduped_in_flight";
-type SessionRefreshTelemetryAttrs = {
-  readonly source: "refreshSessionV2";
-  readonly client_type: AuthSessionClientType;
-  readonly outcome: SessionRefreshTelemetryOutcome;
-  readonly status_code?: number;
-  readonly duration_bucket_ms?: string;
-};
-
-type ApiStatusError = {
-  readonly status?: unknown;
-  readonly response?: {
-    readonly status?: unknown;
-  };
-};
 
 interface CreateConnectionShareResponse {
   readonly connection_share_code: string;
@@ -125,15 +106,6 @@ export function getSessionClientType(): AuthSessionClientType {
   return Capacitor.isNativePlatform() ? "native" : "web";
 }
 
-function isUnauthorizedApiError(error: unknown): boolean {
-  if (typeof error !== "object" || error === null) {
-    return false;
-  }
-
-  const statusError = error as ApiStatusError;
-  return statusError.status === 401 || statusError.response?.status === 401;
-}
-
 function getSessionRefreshKey({
   address,
   clientType,
@@ -153,146 +125,6 @@ const isAbortError = (error: unknown): boolean =>
   error !== null &&
   "name" in error &&
   error.name === "AbortError";
-
-function getSessionRefreshTimingNow(): number {
-  if (globalThis.performance?.now) {
-    return globalThis.performance.now();
-  }
-  return Date.now();
-}
-
-function getKnownApiStatusCode(error: unknown): number | undefined {
-  if (typeof error !== "object" || error === null) {
-    return undefined;
-  }
-
-  const statusError = error as ApiStatusError;
-  const status = statusError.status ?? statusError.response?.status;
-  return typeof status === "number" && Number.isInteger(status)
-    ? status
-    : undefined;
-}
-
-function isNetworkApiError(error: unknown): boolean {
-  if (error instanceof TypeError) {
-    return true;
-  }
-
-  if (!(error instanceof Error)) {
-    return false;
-  }
-
-  const message = error.message.toLowerCase();
-  return (
-    message.includes("network request failed") ||
-    message.includes("network error") ||
-    message.includes("failed to fetch") ||
-    message.includes("load failed")
-  );
-}
-
-function recordSessionRefreshTelemetry(
-  attrs: SessionRefreshTelemetryAttrs
-): void {
-  try {
-    if (
-      attrs.outcome === "backend_error" ||
-      attrs.outcome === "network_error"
-    ) {
-      Sentry.logger.warn("auth_session_refresh", attrs);
-      return;
-    }
-
-    Sentry.logger.info("auth_session_refresh", attrs);
-  } catch {
-    // Telemetry must not affect auth state transitions.
-  }
-}
-
-function recordSessionRefreshOutcome({
-  clientType,
-  outcome,
-  statusCode,
-  startedAtMs,
-}: {
-  readonly clientType: AuthSessionClientType;
-  readonly outcome: SessionRefreshTelemetryOutcome;
-  readonly statusCode?: number | undefined;
-  readonly startedAtMs?: number | undefined;
-}): void {
-  recordSessionRefreshTelemetry({
-    source: "refreshSessionV2",
-    client_type: clientType,
-    outcome,
-    ...(statusCode !== undefined ? { status_code: statusCode } : {}),
-    ...(startedAtMs !== undefined
-      ? {
-          duration_bucket_ms: bucketMs(
-            getSessionRefreshTimingNow() - startedAtMs
-          ),
-        }
-      : {}),
-  });
-}
-
-function recordSessionRefreshFailureOutcome({
-  clientType,
-  error,
-  startedAtMs,
-}: {
-  readonly clientType: AuthSessionClientType;
-  readonly error: unknown;
-  readonly startedAtMs: number;
-}): void {
-  const statusCode = getKnownApiStatusCode(error);
-  if (statusCode === 401) {
-    recordSessionRefreshOutcome({
-      clientType,
-      outcome: "unauthorized",
-      statusCode,
-      startedAtMs,
-    });
-    return;
-  }
-
-  if (statusCode !== undefined) {
-    recordSessionRefreshOutcome({
-      clientType,
-      outcome: "backend_error",
-      statusCode,
-      startedAtMs,
-    });
-    return;
-  }
-
-  if (isNetworkApiError(error)) {
-    recordSessionRefreshOutcome({
-      clientType,
-      outcome: "network_error",
-      startedAtMs,
-    });
-  }
-}
-
-async function withSessionRefreshAbortTelemetry<T>({
-  clientType,
-  task,
-}: {
-  readonly clientType: AuthSessionClientType;
-  readonly task: () => Promise<T>;
-}): Promise<T> {
-  try {
-    return await task();
-  } catch (error: unknown) {
-    if (isAbortError(error)) {
-      recordSessionRefreshOutcome({
-        clientType,
-        outcome: "aborted",
-      });
-    }
-    throw error;
-  }
-}
 
 function getActiveFailureCooldown(
   key: string
@@ -526,6 +358,10 @@ async function executeSessionRefreshV2({
   if (clientType !== "web") {
     const nativeRefreshToken = await getNativeRefreshToken(address);
     if (!nativeRefreshToken) {
+      recordSessionRefreshOutcome({
+        clientType,
+        outcome: "unauthorized",
+      });
       return null;
     }
 
@@ -584,7 +420,7 @@ async function executeSessionRefreshRequest<T extends SessionRefreshResponse>({
   readonly clientType: AuthSessionClientType;
   readonly request: () => Promise<T>;
 }): Promise<T | null> {
-  const startedAtMs = getSessionRefreshTimingNow();
+  const startedAtMs = getSessionRefreshTelemetryTimestamp();
   recordSessionRefreshOutcome({
     clientType,
     outcome: "started",
@@ -605,7 +441,7 @@ async function executeSessionRefreshRequest<T extends SessionRefreshResponse>({
       startedAtMs,
     });
 
-    if (isUnauthorizedApiError(error)) {
+    if (isUnauthorizedSessionRefreshError(error)) {
       return null;
     }
     throw error;
